# docs/openapi.internal.v1.yaml
openapi: 3.0.3
info:
  title: Jarvis Internal Application API
  version: 1.0.0
  description: >
    Internal control/application API for Jarvis Core. Not a public customer API.
    Measurement submission is an explicit trusted server-side operation only.
servers:
  - url: /
    description: Same origin as the Jarvis process
security:
  - bearerAuth: []
paths:
  /health:
    get:
      security: []
      summary: Public health check
      operationId: getHealth
      responses:
        '200':
          description: Service is up
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [status, service]
                properties:
                  status:
                    type: string
                    enum: [ok]
                  service:
                    type: string
                    enum: [supermoskitka-jarvis]

  /internal/v1/conversations:
    post:
      summary: Create conversation
      operationId: createConversation
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                channel:
                  type: string
                  description: Optional channel label (e.g. telegram)
                customerId:
                  type: string
                  description: Optional external customer identifier
      responses:
        '201':
          description: Conversation created (does not create CRM leads or measurements)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/NotConfigured'

  /internal/v1/conversations/{conversationId}:
    get:
      summary: Get conversation
      operationId: getConversation
      parameters:
        - $ref: '#/components/parameters/ConversationId'
      responses:
        '200':
          description: Conversation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/NotConfigured'

  /internal/v1/conversations/{conversationId}/messages:
    get:
      summary: List transcript messages
      operationId: listConversationMessages
      parameters:
        - $ref: '#/components/parameters/ConversationId'
      responses:
        '200':
          description: Messages in chronological order
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [messages]
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/MessageDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/NotConfigured'
    post:
      summary: Ingress customer message
      operationId: handleCustomerMessage
      description: >
        Persists the customer message and, when mode is AI, runs the application
        turn. Idempotent for the same conversationId + messageId + text
        (duplicate=true). Same messageId with different text returns 409.
      parameters:
        - $ref: '#/components/parameters/ConversationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [messageId, text]
              properties:
                messageId:
                  type: string
                  description: Channel-stable message id used for idempotency
                text:
                  type: string
                createdAt:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Processed or idempotent duplicate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandleCustomerMessageResultDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Message id conflict or persistence conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Upstream LLM/provider unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          $ref: '#/components/responses/NotConfigured'
        '500':
          $ref: '#/components/responses/InternalError'

  /internal/v1/conversations/{conversationId}/mode:
    post:
      summary: Set AI/HUMAN mode
      operationId: setConversationMode
      description: >
        Switches ownership without clearing transcript, Order Memory, or quote.
        HUMAN mode persists customer messages but does not call FactExtractor / LLM.
      parameters:
        - $ref: '#/components/parameters/ConversationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [mode]
              properties:
                mode:
                  type: string
                  enum: [AI, HUMAN]
      responses:
        '200':
          description: Mode updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationDto'
        '400':
          description: Invalid mode (MODE_INVALID / VALIDATION_ERROR)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/NotConfigured'

  /internal/v1/conversations/{conversationId}/order-state:
    get:
      summary: Internal order/lead state
      operationId: getConversationOrderState
      description: >
        Compact internal DTO for future CRM adapters. Never returns supplier/BOM
        dumps, trusted proof internals, system prompts, or measurement payouts.
      parameters:
        - $ref: '#/components/parameters/ConversationId'
      responses:
        '200':
          description: Compact internal order-state DTO
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationOrderStateDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/NotConfigured'

  /internal/v1/conversations/{conversationId}/measurement-action:
    get:
      summary: Measurement action decision (+ safe draft when ready)
      operationId: getMeasurementAction
      description: >
        Uses readiness + measurement action policy. Reading AUTO_ALLOWED never
        writes operational data.
      parameters:
        - $ref: '#/components/parameters/ConversationId'
      responses:
        '200':
          description: Measurement action decision (optional safe draft when not NOT_READY)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasurementActionDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/NotConfigured'

  /internal/v1/conversations/{conversationId}/measurement-submit:
    post:
      summary: Explicitly submit the current trusted measurement draft
      operationId: submitReadyMeasurement
      description: >
        Rebuilds all values from the current server-side Order Memory and trusted
        preliminary quote. Request-body phone, address, items, total, or quote
        overrides are ignored. This operation is never triggered automatically
        by an ordinary customer message.
      parameters:
        - $ref: '#/components/parameters/ConversationId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              description: No request-body fields are used as measurement authority.
      responses:
        '200':
          description: Firestore upsert and Sheet projection succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasurementSubmitResultDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: >
            Owner approval is required or the memory/quote changed before execution
            (MEASUREMENT_OWNER_APPROVAL_REQUIRED / MEASUREMENT_SUBMISSION_STALE).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Current MeasurementAction is NOT_READY (MEASUREMENT_NOT_READY)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: >
            Firestore persistence failed, or Firestore was upserted and the Sheet
            projection failed. Partial details identify firestore=UPSERTED.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: >
            Internal API runtime or measurement Sheet webhook is not configured.
            MEASUREMENT_SHEET_NOT_CONFIGURED includes partial Firestore status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Shared JARVIS_INTERNAL_API_KEY bearer token

  parameters:
    ConversationId:
      name: conversationId
      in: path
      required: true
      schema:
        type: string

  responses:
    Unauthorized:
      description: Missing or invalid bearer token (UNAUTHORIZED)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotConfigured:
      description: JARVIS_INTERNAL_API_KEY or application runtime is not configured (INTERNAL_API_NOT_CONFIGURED)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Conversation not found (CONVERSATION_NOT_FOUND)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Validation or invalid input (VALIDATION_ERROR / MODE_INVALID)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Unexpected internal error (INTERNAL_ERROR)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    ConversationDto:
      type: object
      additionalProperties: false
      required: [conversationId, mode, channel, createdAt, updatedAt]
      properties:
        conversationId:
          type: string
        mode:
          type: string
          enum: [AI, HUMAN]
        channel:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string

    MessageDto:
      type: object
      additionalProperties: false
      required: [messageId, sender, text, createdAt]
      properties:
        messageId:
          type: string
        sender:
          type: string
          enum: [CUSTOMER, AI, HUMAN]
        text:
          type: string
        createdAt:
          type: string

    AiReplyDto:
      type: object
      additionalProperties: false
      required: [messageId, text]
      properties:
        messageId:
          type: string
        text:
          type: string

    HandleCustomerMessageResultDto:
      type: object
      additionalProperties: false
      required: [conversationId, conversationMode, customerMessageId, duplicate, aiReply]
      properties:
        conversationId:
          type: string
        conversationMode:
          type: string
          enum: [AI, HUMAN]
        customerMessageId:
          type: string
        duplicate:
          type: boolean
          description: True when the same conversationId + messageId + text was already processed
        resumed:
          type: boolean
          description: True when a previously persisted incomplete customer turn was resumed
        aiReply:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/AiReplyDto'
          description: Null in HUMAN mode or when no AI reply was produced

    CustomerStateDto:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        customerType:
          type: string

    OrderItemStateDto:
      type: object
      additionalProperties: false
      required: [localItemId]
      properties:
        localItemId:
          type: string
        productType:
          type: string
        quantity:
          type: number
        widthMm:
          type: number
        heightMm:
          type: number
        measurementBasis:
          type: string
        mesh:
          type: string
        profile:
          type: string
        profileColor:
          type: string
        ral:
          type: string
        finish:
          type: string
        fastening:
          type: string
        opening:
          type: string
        comment:
          type: string

    PreliminaryQuoteStateDto:
      type: object
      additionalProperties: false
      required: [quoteId, publicTotalRub, current, accepted]
      properties:
        quoteId:
          type: string
        publicTotalRub:
          type: number
        current:
          type: boolean
        accepted:
          type: boolean

    ReadinessDto:
      type: object
      additionalProperties: false
      required: [status, missingCodes]
      properties:
        status:
          type: string
          enum: [NOT_READY, READY_FOR_MEASUREMENT]
        missingCodes:
          type: array
          items:
            type: string
            enum:
              - NEEDS_INPUT
              - NEEDS_SIZE_BASIS
              - PRODUCT_MISSING
              - PRICE_NOT_ACCEPTED
              - MEASUREMENT_NOT_AGREED
              - QUOTE_MISSING
              - QUOTE_STALE
              - PRICING_POLICY_INCOMPLETE
              - CONTACT_MISSING
              - ADDRESS_MISSING

    MeasurementActionSummaryDto:
      type: object
      additionalProperties: false
      required: [kind]
      properties:
        kind:
          type: string
          enum: [NOT_READY, AUTO_ALLOWED, AWAITING_OWNER_APPROVAL]

    ProfitabilitySummaryDto:
      type: object
      additionalProperties: false
      required: [costBasisStatus, profitabilityBand]
      properties:
        costBasisStatus:
          type: string
          enum: [EXACT, PARTIAL, UNAVAILABLE]
        grossProfitRub:
          type: number
          description: Present only when costBasisStatus is EXACT
        grossMarginPercent:
          type: number
          description: Present only when costBasisStatus is EXACT
        markupPercent:
          type: number
          description: Present only when costBasisStatus is EXACT
        profitabilityBand:
          type: string
          enum: [GREEN, YELLOW, RED, UNAVAILABLE]

    ConversationOrderStateDto:
      type: object
      additionalProperties: false
      required:
        - conversationId
        - memoryRevision
        - customer
        - items
        - readiness
        - measurementAction
      properties:
        conversationId:
          type: string
        memoryRevision:
          type: integer
        customer:
          $ref: '#/components/schemas/CustomerStateDto'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemStateDto'
        preliminaryQuote:
          $ref: '#/components/schemas/PreliminaryQuoteStateDto'
        measurementAgreed:
          type: boolean
        readiness:
          $ref: '#/components/schemas/ReadinessDto'
        measurementAction:
          $ref: '#/components/schemas/MeasurementActionSummaryDto'
        profitability:
          $ref: '#/components/schemas/ProfitabilitySummaryDto'

    FulfillmentDraftDto:
      type: object
      additionalProperties: false
      properties:
        installationRequested:
          type: boolean
        pickupRequested:
          type: boolean
        deliveryRequested:
          type: boolean
        deliveryType:
          type: string
        deliveryKm:
          type: number

    MeasurementDraftItemDto:
      type: object
      additionalProperties: false
      required: [localItemId]
      properties:
        localItemId:
          type: string
        productType:
          type: string
        quantity:
          type: number
        widthMm:
          type: number
        heightMm:
          type: number
        measurementBasis:
          type: string
        mesh:
          type: string
        profile:
          type: string
        profileColor:
          type: string
        ral:
          type: string
        finish:
          type: string
        fastening:
          type: string
        opening:
          type: string

    MeasurementDraftQuoteDto:
      type: object
      additionalProperties: false
      required: [quoteId, publicTotalRub]
      properties:
        quoteId:
          type: string
        publicTotalRub:
          type: number

    MeasurementDraftCustomerDto:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string

    MeasurementDraftDto:
      type: object
      additionalProperties: false
      required: [conversationId, memoryRevision, customer, items, fulfillment]
      properties:
        conversationId:
          type: string
        memoryRevision:
          type: integer
        customer:
          $ref: '#/components/schemas/MeasurementDraftCustomerDto'
        items:
          type: array
          items:
            $ref: '#/components/schemas/MeasurementDraftItemDto'
        fulfillment:
          $ref: '#/components/schemas/FulfillmentDraftDto'
        preliminaryQuote:
          $ref: '#/components/schemas/MeasurementDraftQuoteDto'

    MeasurementActionDto:
      type: object
      additionalProperties: false
      required: [conversationId, kind, readiness]
      properties:
        conversationId:
          type: string
        kind:
          type: string
          enum: [NOT_READY, AUTO_ALLOWED, AWAITING_OWNER_APPROVAL]
        readiness:
          $ref: '#/components/schemas/ReadinessDto'
        draft:
          description: Safe draft when kind is not NOT_READY; no internal economics
          allOf:
            - $ref: '#/components/schemas/MeasurementDraftDto'

    MeasurementSubmitResultDto:
      type: object
      additionalProperties: false
      required: [conversationId, submissionId, status, firestore, sheet]
      properties:
        conversationId:
          type: string
        submissionId:
          type: string
          description: Deterministic jarvis_<SHA-256 prefix> idempotency key
        status:
          type: string
          enum: [SUBMITTED]
        firestore:
          type: string
          enum: [UPSERTED]
        sheet:
          type: string
          enum: [SENT]

    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId]
          properties:
            code:
              type: string
              enum:
                - UNAUTHORIZED
                - INTERNAL_API_NOT_CONFIGURED
                - VALIDATION_ERROR
                - CONVERSATION_NOT_FOUND
                - MESSAGE_ID_CONFLICT
                - DUPLICATE_MESSAGE
                - MODE_INVALID
                - PERSISTENCE_CONFLICT
                - PROVIDER_UNAVAILABLE
                - MEASUREMENT_NOT_READY
                - MEASUREMENT_OWNER_APPROVAL_REQUIRED
                - MEASUREMENT_SUBMISSION_STALE
                - MEASUREMENT_SHEET_NOT_CONFIGURED
                - MEASUREMENT_SHEET_FAILED
                - MEASUREMENT_PERSISTENCE_FAILED
                - INTERNAL_ERROR
            message:
              type: string
            requestId:
              type: string
            details:
              type: object
              additionalProperties: true
              description: >
                Safe operational status only. Partial measurement failures may
                include conversationId, submissionId, status, firestore, and sheet.